{"schema":"decodex/commit/1","summary":"Modularize lifecycle hook and lane control surfaces","authority":"manual"}
